+++ src/knowledge-memory/contracts/integration-interoperability-profile.js
@@ -1,0 +1,62 @@
+export class KnowledgeMemoryIntegrationProfile {
+  constructor({
+    profileName,
+    objectives = [], participants = [], profileFields = [], integrationStyles = [],
+    envelopeFields = [], apiCapabilities = [], apiContractFields = [], responseStatuses = [],
+    errorFields = [], eventEnvelopeFields = [], eventDeliveryControls = [], connectorCapabilities = [],
+    checkpointFields = [], bulkManifestFields = [], bulkItemStatuses = [], portabilityFields = [],
+    degradedModes = [], qualityAttributes = [], architecturalRules = [], architectureBoundaries = [],
+    publishedContracts = true, authoritativeOwnership = true, semanticCompatibility = true,
+    capabilityEncapsulation = true, providerAbstraction = true, identityPropagated = true,
+    purposePropagated = true, trustedScopePropagated = true, classificationPropagated = true,
+    provenancePropagated = true, temporalSemantics = true, referenceFirst = true,
+    referenceAuthorizationIndependent = true, statusSemanticsDistinct = true,
+    safeErrorContracts = true, idempotentMutations = true, concurrencyControlled = true,
+    deadlinesBounded = true, contractsVersioned = true, backwardCompatible = true,
+    deprecationGoverned = true, immutableEvents = true, atLeastOnceSafe = true,
+    replaySafe = true, deadLetterGoverned = true, reconciliationEnabled = true,
+    schemasGoverned = true, connectorsGoverned = true, antiCorruptionLayers = true,
+    domainTruthPreserved = true, externalOutputValidated = true, bulkControlParity = true,
+    portableLifecycle = true, correctionPropagated = true, deletionPropagated = true,
+    projectionsSynchronized = true, divergenceContained = true, isolationEndToEnd = true,
+    secretsExcluded = true, evidenceProtected = true, safeDegradation = true,
+    vendorNeutral = true, technologyIndependent = true,
+    directDatabaseAccess = false, sharedInternalTables = false, hiddenFilesystemExchange = false,
+    unversionedPayloads = false, providerObjectsCanonical = false, searchIndexAuthoritative = false,
+    copiedCredentialsGrantAuthority = false, transportCreatesAuthority = false,
+    payloadTextDefinesScope = false, referenceGrantsAccess = false, classificationLoweredByTransform = false,
+    acceptanceMeansPublication = false, providerOutputAutoApproved = false,
+    connectorPublishesKnowledge = false, eventIsUnboundedCommand = false, exactlyOnceAssumed = false,
+    replayResurrectsRecords = false, bulkWeakensControls = false, integrationSharesSecrets = false,
+    degradedWeakensControls = false, selectsIntegrationProduct = false
+  }) {
+    this.profileName = profileName;
+    for (const [key, value] of Object.entries({
+      objectives, participants, profileFields, integrationStyles, envelopeFields, apiCapabilities,
+      apiContractFields, responseStatuses, errorFields, eventEnvelopeFields, eventDeliveryControls,
+      connectorCapabilities, checkpointFields, bulkManifestFields, bulkItemStatuses, portabilityFields,
+      degradedModes, qualityAttributes, architecturalRules, architectureBoundaries
+    })) this[key] = Object.freeze([...value]);
+    for (const [key, value] of Object.entries({
+      publishedContracts, authoritativeOwnership, semanticCompatibility, capabilityEncapsulation,
+      providerAbstraction, identityPropagated, purposePropagated, trustedScopePropagated,
+      classificationPropagated, provenancePropagated, temporalSemantics, referenceFirst,
+      referenceAuthorizationIndependent, statusSemanticsDistinct, safeErrorContracts,
+      idempotentMutations, concurrencyControlled, deadlinesBounded, contractsVersioned,
+      backwardCompatible, deprecationGoverned, immutableEvents, atLeastOnceSafe, replaySafe,
+      deadLetterGoverned, reconciliationEnabled, schemasGoverned, connectorsGoverned,
+      antiCorruptionLayers, domainTruthPreserved, externalOutputValidated, bulkControlParity,
+      portableLifecycle, correctionPropagated, deletionPropagated, projectionsSynchronized,
+      divergenceContained, isolationEndToEnd, secretsExcluded, evidenceProtected,
+      safeDegradation, vendorNeutral, technologyIndependent, directDatabaseAccess,
+      sharedInternalTables, hiddenFilesystemExchange, unversionedPayloads,
+      providerObjectsCanonical, searchIndexAuthoritative, copiedCredentialsGrantAuthority,
+      transportCreatesAuthority, payloadTextDefinesScope, referenceGrantsAccess,
+      classificationLoweredByTransform, acceptanceMeansPublication, providerOutputAutoApproved,
+      connectorPublishesKnowledge, eventIsUnboundedCommand, exactlyOnceAssumed,
+      replayResurrectsRecords, bulkWeakensControls, integrationSharesSecrets,
+      degradedWeakensControls, selectsIntegrationProduct
+    })) this[key] = Boolean(value);
+    Object.freeze(this);
+  }
+}